requirements.txt
Django==6.0.3
Pillow==12.1.1
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
gunicorn==23.0.0
whitenoise[brotli]==6.9.0
dj-database-url==2.3.0
